Fix issues rebasing onto latest alpha-dev
