+++ Dependencies/MelonStartScreen/UI/Themes/UI_Theme_Default.cs
@@ -5,9 +5,9 @@
         internal UI_Theme_Default() => Defaults();
 
         internal override byte[] GetLoadingImage()
-            => Properties.Resources.Loading_Melon;
+            => StartScreenResources.MelonLoadingIcon;
 
         internal override byte[] GetLogoImage()
-            => Properties.Resources.Logo_Melon;
+            => StartScreenResources.MelonLogo;
     }
 }
