+++ Dependencies/MelonStartScreen/UI/Themes/UI_Theme_Lemon.cs
@@ -5,9 +5,9 @@
         internal UI_Theme_Lemon() => Defaults();
 
         internal override byte[] GetLoadingImage()
-            => Properties.Resources.Loading_Lemon;
+            => StartScreenResources.LemonLoadingIcon;
 
         internal override byte[] GetLogoImage()
-            => Properties.Resources.Logo_Lemon;
+            => StartScreenResources.LemonLogo;
     }
 }
