+++ MelonLoader/InternalUtils/UnityInformationHandler.cs
@@ -3,17 +3,12 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.RegularExpressions;
 using AssetsTools.NET;
 using AssetsTools.NET.Extra;
 using UnityVersion = AssetRipper.VersionUtilities.UnityVersion;
-using System.Diagnostics;
 using System.Drawing;
-using System.Text;
-using System.Text.RegularExpressions;
 using MelonLoader.Utils;
 
 namespace MelonLoader.InternalUtils
@@ -143,7 +138,7 @@ namespace MelonLoader.InternalUtils
         {
             try
             {
-                string appInfoFilePath = Path.Combine(MelonUtils.GetGameDataDirectory(), "app.info");
+                string appInfoFilePath = Path.Combine(MelonEnvironment.UnityGameDataDirectory, "app.info");
                 if (!File.Exists(appInfoFilePath))
                     return;
